Rename callAssignment store/actions to "assignment"
There is no doubt in the "call.zetk.in" app that assignments are call
assignments, and not some other, future kind of assignments.
